test: split test_threadsafe_function to avoid timeouts
test_threadsafe_function times out sometimes on some platforms. (When it
passes, it takes nearly a minute on some of our ARM machines.)
